// types.ts
export interface CompanyInfo {
  company_name: string;
  furigana: string | null;
  address: string;
  corporate_number: string;
  representative_name: string | null;
  industry: string | null;
  url: string;
  founded_date: string | null;
  capital: string | null;
  employee_count: string | null;
  phone_number: string | null;
  listing_status: string | null;
  stock_code: string | null;
  invoice_registration_number: string | null;
  introduction: string | null;
}